move data types into the datatypes package
