+++ charts/geo.go
@@ -2,6 +2,7 @@ package charts
 
 import (
 	"fmt"
+	"github.com/chenjiandongx/go-echarts/datatypes"
 	"io"
 	"log"
 
@@ -43,9 +44,9 @@ func NewGeo(mapType string, routers ...RouterOpts) *Geo {
 // common.ChartType.EffectScatter
 // common.ChartType.HeatMap
 func (c *Geo) Add(name, geoType string, data map[string]float32, options ...seriesOptser) *Geo {
-	nvs := make([]common.NameValueItem, 0)
+	nvs := make([]datatypes.NameValueItem, 0)
 	for k, v := range data {
-		nvs = append(nvs, common.NameValueItem{Name: k, Value: c.extendValue(k, v)})
+		nvs = append(nvs, datatypes.NameValueItem{Name: k, Value: c.extendValue(k, v)})
 	}
 	series := singleSeries{Name: name, Type: geoType, Data: nvs, CoordSystem: common.ChartType.Geo}
 	series.setSingleSeriesOpts(options...)
