+++ charts/wordcloud.go
@@ -1,6 +1,7 @@
 package charts
 
 import (
+	"github.com/chenjiandongx/go-echarts/datatypes"
 	"io"
 
 	"github.com/chenjiandongx/go-echarts/common"
@@ -50,9 +51,9 @@ func NewWordCloud(routers ...RouterOpts) *WordCloud {
 
 // Add adds new data sets.
 func (c *WordCloud) Add(name string, data map[string]interface{}, options ...seriesOptser) *WordCloud {
-	nvs := make([]common.NameValueItem, 0)
+	nvs := make([]datatypes.NameValueItem, 0)
 	for k, v := range data {
-		nvs = append(nvs, common.NameValueItem{Name: k, Value: v})
+		nvs = append(nvs, datatypes.NameValueItem{Name: k, Value: v})
 	}
 	series := singleSeries{Name: name, Type: common.ChartType.WordCloud, Data: nvs}
 	series.setSingleSeriesOpts(options...)
